+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### ADDED
 
+- A callback function after each disconnection.
 - Added support for SDM: 
   - Added support for multi-core and multi-mode for Link object
   - Added Test cases in test_link.cpp for the corresponding methods of multi-core/mode Link object.
@@ -23,6 +24,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added getters for BitRate class:
   - getNumberOfModulations() gets the number of modulations available in the BitRate object.
   - Added tests.
+
+### FIXED
+
+- Example 001 fixed. Reset slots in each route.
 
 ## [0.5.3] - 2022-05-19